Fixed title bug
Was triggering a treeview title update when programmatically changing the title when open a new note.
We open a new note, and at the same time as putting the opening note's title in we replaced it with the old one (because we detected a title change).

Changed it the title update callback to trigger from user only changes, textEdited signal instead of textChanged